Checkout 'all' projects for forestdb
Forestdb isn't part of the 'kv' group or similar; for us to be able to
test it we need to also check out the 'forestdb' repo.

However, that also requires that we don't set
COUCHBASE_KV_COMMIT_VALIDATION (see f5e7d07) as otherwise CMake will
ignore forestdb. However, if COUCHBASE_KV_COMMIT_VALIDATION is not set
then CMake requires _all_ repo projects to build correctly.

As such, modify the behavour of our repo checkout to checkout all
projects if the tested project is 'forestdb'.

Change-Id: I8542b1880781949bddcb31e18a4d09e0f0bc9e5e
Reviewed-on: http://review.couchbase.org/c/server-cv/+/159293
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>