CBD-6071: Remove verify-status actions
The verify-status plugin for Jenkins has not been updated to work with
new releases of Gerrit. We will no longer report back to gerrit this
way, but via comments (and checks-jenkins/Gerrit Checks once an
outstanding authentication issue has been ironed out)

This change also disables Gerrit Trigger's silentStartMode in the
interim to avoid confusion, so that a comment will be posted on the
change detailing that the CV job has begun.

Change-Id: I84aec46fe6dc6823426f24b9ae7c0b53aaafe98a
Reviewed-on: https://review.couchbase.org/c/server-cv/+/218207
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Blair Watt <blair.watt@couchbase.com>