CBD-6071: Remove verify-status actions
The verify-status plugin for Jenkins has not been updated to work with
new releases of Gerrit. We will no longer report back to gerrit this
way, but via comments (and checks-jenkins/Gerrit Checks once an
outstanding authentication issue has been ironed out)

This change also disables Gerrit Trigger's silentStartMode in the
interim to avoid confusion, so that a comment will be posted on the
change detailing that the CV job has begun.

Change-Id: I441a1af367f67f66366e65ee81c86ca58e5d8702
Reviewed-on: https://review.couchbase.org/c/server-cv/+/219315
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Blair Watt <blair.watt@couchbase.com>