CBD-6071: Remove verify-status actions
The verify-status plugin for Jenkins has not been updated to work with
new releases of Gerrit. We will no longer report back to gerrit this
way, but via comments (and checks-jenkins/Gerrit Checks once an
outstanding authentication issue has been ironed out)

This change also disables Gerrit Triger's silentStartMode in the
interim to avoid confusion, so that a comment will be posted on the
change detailing that the CV job has begun.

Change-Id: I6b6dcc7a53450fac8d15a2991d50f733d39a42b8
Reviewed-on: https://review.couchbase.org/c/server-cv/+/219808
Tested-by: Blair Watt <blair.watt@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>