CBD-6071: Remove verify-status actions
The verify-status plugin for Jenkins has not been updated to work with
new releases of Gerrit. We will no longer report back to gerrit this
way, but via comments (and checks-jenkins/Gerrit Checks once an
outstanding authentication issue has been ironed out)

This change also disables Gerrit Triger's silentStartMode in the
interim to avoid confusion, so that a comment will be posted on the
change detailing that the CV job has begun.

Change-Id: I630aef4848206a6ff9d02cd3430d132d95e3dc67
Reviewed-on: https://review.couchbase.org/c/server-cv/+/220098
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Blair Watt <blair.watt@couchbase.com>